pyproject.toml
[project]
name = "agent-sierra"
version = "0.1.0"
description = "Sentry webhook → Cursor Cloud Agent for auto-fixing unhandled errors"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
]
